+++ lib/rig_api/v1/messages_test.exs
@@ -1,4 +1,4 @@
-defmodule RigApi.MessageControllerTest do
+defmodule RigApi.V1.MessagesTest do
   @moduledoc false
   use RigApi.ConnCase, async: true