+++ lib/rig_api/v2/responses.ex
@@ -1,4 +1,5 @@
-defmodule RigApi.ResponsesController do
+defmodule RigApi.V2.Responses do
+  @moduledoc "Controller for submitting (backend) responses to asynchronous (frontend) requests."
   require Logger
 
   use RigApi, :controller
@@ -7,10 +8,12 @@ defmodule RigApi.ResponsesController do
   alias Rig.Connection.Codec
   alias RigCloudEvents.CloudEvent
 
-  action_fallback(RigApi.FallbackController)
+  @prefix "/v2"
+
+  action_fallback(RigApi.Fallback)
 
   swagger_path :create do
-    post("/v1/responses")
+    post(@prefix <> "/responses")
     summary("Submit a message, to be sent to correlated reverse proxy request.")
     description("Allows you to submit a message to RIG using a simple, \
     synchronous call. Message will be sent to correlated reverse proxy request.")