+++ lib/rig_inbound_gateway_web/router.ex
@@ -12,33 +12,37 @@ defmodule RigInboundGatewayWeb.Router do
     get("/health", HealthController, :check_health)
 
     scope "/v1", V1 do
-      scope "/connection/sse" do
-        subscription_url = "/:connection_id/subscriptions"
-        options(subscription_url, SubscriptionController, :handle_preflight)
-        put(subscription_url, SubscriptionController, :set_subscriptions)
+      scope "/connection" do
+        get("/init", ConnectionController, :init)
 
-        # The SSE handler is implemented using Cowboy's loop handler behaviour and set
-        # up using the Cowboy dispatch configuration; see the `config.exs` file.
+        scope "/sse" do
+          subscription_url = "/:connection_id/subscriptions"
+          options(subscription_url, SubscriptionController, :handle_preflight)
+          put(subscription_url, SubscriptionController, :set_subscriptions)
+          
+          metadata_url = "/:connection_id/metadata"
+          put(metadata_url, MetadataController, :set_metadata)
 
-        metadata_url = "/:connection_id/metadata"
-        put(metadata_url, MetadataController, :set_metadata)
-      end
+          # The SSE handler is implemented using Cowboy's loop handler behaviour and set
+          # up using the Cowboy dispatch configuration; see the `config.exs` file.
+        end
 
-      # Unlike SSE & WS handlers, the LP handler is implemented using plug
-      scope "/connection/longpolling" do
-        subscription_url = "/:connection_id/subscriptions"
-        get("/", LongpollingController, :handle_connection)
-        options(subscription_url, SubscriptionController, :handle_preflight)
-        put(subscription_url, SubscriptionController, :set_subscriptions)
-      end
+        scope "/ws" do
+          subscription_url = "/:connection_id/subscriptions"
+          options(subscription_url, SubscriptionController, :handle_preflight)
+          put(subscription_url, SubscriptionController, :set_subscriptions)
 
-      scope "/connection/ws" do
-        subscription_url = "/:connection_id/subscriptions"
-        options(subscription_url, SubscriptionController, :handle_preflight)
-        put(subscription_url, SubscriptionController, :set_subscriptions)
+          # The WebSocket handler is implemented using Cowboy's loop handler behaviour and set
+          # up using the Cowboy dispatch configuration; see the `config.exs` file.
+        end
 
-        # The WebSocket handler is implemented using Cowboy's loop handler behaviour and set
-        # up using the Cowboy dispatch configuration; see the `config.exs` file.
+        # Unlike SSE & WS handlers, the LP handler is implemented using plug
+        scope "/longpolling" do
+          subscription_url = "/:connection_id/subscriptions"
+          get("/", LongpollingController, :handle_connection)
+          options(subscription_url, SubscriptionController, :handle_preflight)
+          put(subscription_url, SubscriptionController, :set_subscriptions)
+        end
       end
 
       options("/events", EventController, :handle_preflight)