+++ test/rig_tests/proxy/request_logger/kafka_test.exs
@@ -67,7 +67,7 @@ defmodule RigTests.Proxy.RequestLogger.KafkaTest do
     endpoint_path = "/#{endpoint_id}"
 
     # We register the endpoint with the proxy:
-    rig_api_url = "http://localhost:#{@api_port}/v1/apis"
+    rig_api_url = "http://localhost:#{@api_port}/v2/apis"
     rig_proxy_url = "http://localhost:#{@proxy_port}"
 
     route(endpoint_path, Response.ok!(~s<{"status":"ok"}>))
@@ -134,7 +134,7 @@ defmodule RigTests.Proxy.RequestLogger.KafkaTest do
     endpoint_path = "/#{endpoint_id}"
 
     # We register the endpoint with the proxy:
-    rig_api_url = "http://localhost:#{@api_port}/v1/apis"
+    rig_api_url = "http://localhost:#{@api_port}/v2/apis"
     rig_proxy_url = "http://localhost:#{@proxy_port}"
 
     route(endpoint_path, Response.ok!(~s<{"status":"ok"}>))