+++ test/support/clients.ex
@@ -1,14 +1,3 @@
-defmodule Client do
-  @moduledoc false
-  @type client :: pid | reference | atom
-  @callback connect(params :: list) :: {:ok, pid}
-  @callback disconnect(client) :: :ok
-  @callback refute_receive(client) :: :ok
-  @callback read_event(client, event_type :: String.t()) :: map()
-  @callback read_welcome_event(client) :: map()
-  @callback read_subscriptions_set_event(client) :: map()
-end
-
 defmodule TestClient.ConnectionError do
   defexception [:code, :reason]
 
@@ -18,6 +7,20 @@ defmodule TestClient.ConnectionError do
   def message(%__MODULE__{code: code, reason: reason}),
     do:
       "could not establish connection, server responded with #{inspect(code)}: #{inspect(reason)}"
+end
+
+defmodule Client do
+  @moduledoc false
+  @type client :: pid | reference | atom | map
+  @callback connect(params :: list) :: {:ok, pid}
+  @callback disconnect(client) :: :ok
+  @callback status(client) :: {:open, client} | {:closed, client}
+  @callback refute_receive(client) :: :ok
+  @callback read_event(client, event_type :: String.t()) :: map()
+  @callback read_welcome_event(client) :: map()
+  @callback read_subscriptions_set_event(client) :: map()
+  @callback try_connect_then_disconnect(params :: list) ::
+              {:ok, status :: any} | {:error, reason :: %TestClient.ConnectionError{}}
 end
 
 defmodule SseClient do
@@ -70,27 +73,39 @@ defmodule SseClient do
       500 -> raise "No response"
     end
 
-    {:ok, client}
+    {:ok, %{client: client, status: :open}}
   end
 
   @impl true
-  def disconnect(client) do
+  def disconnect(%{client: client, status: :open}) do
     {:ok, ^client} = :hackney.stop_async(client)
     :ok
   end
 
   @impl true
-  def refute_receive(ignored_client_ref) do
+  def status(state)
+  def status(%{status: :closed} = state), do: {:closed, state}
+
+  def status(%{client: client} = state) do
     receive do
-      %HTTPoison.AsyncChunk{} = async_chunk ->
-        raise "Unexpectedly received: #{inspect(async_chunk)}"
+      %HTTPoison.AsyncEnd{id: ^client} -> {:closed, %{state | status: :closed}}
     after
-      100 -> {:ok, ignored_client_ref}
+      100 -> {:open, state}
     end
   end
 
   @impl true
-  def read_event(ignored_client_ref, event_type) do
+  def refute_receive(%{status: :open} = state) do
+    receive do
+      %HTTPoison.AsyncChunk{} = async_chunk ->
+        raise "Unexpectedly received: #{inspect(async_chunk)}"
+    after
+      100 -> {:ok, state}
+    end
+  end
+
+  @impl true
+  def read_event(%{status: :open} = state, event_type) do
     cloud_event =
       read_sse_chunk()
       |> extract_cloud_event()
@@ -101,14 +116,26 @@ defmodule SseClient do
         %{"cloudEventsVersion" => "0.1", "eventType" => ^event_type} -> cloud_event
       end
 
-    {cloud_event, ignored_client_ref}
+    {cloud_event, state}
   end
 
   @impl true
-  def read_welcome_event(client), do: read_event(client, "rig.connection.create")
+  def read_welcome_event(state), do: read_event(state, "rig.connection.create")
 
   @impl true
-  def read_subscriptions_set_event(client), do: read_event(client, "rig.subscriptions_set")
+  def read_subscriptions_set_event(state), do: read_event(state, "rig.subscriptions_set")
+
+  @impl true
+  def try_connect_then_disconnect(params \\ []) do
+    {:ok, client} = connect(params)
+    flush_mailbox()
+    disconnect(client)
+    {:ok, client}
+  rescue
+    err in [TestClient.ConnectionError] -> {:error, err}
+  after
+    flush_mailbox()
+  end
 
   defp read_sse_chunk do
     receive do
@@ -134,7 +161,7 @@ defmodule SseClient do
     end
   end
 
-  defp flush_mailbox do
+  def flush_mailbox do
     receive do
       msg -> [msg | flush_mailbox()]
     after
@@ -186,25 +213,50 @@ defmodule WsClient do
           message
       end
 
-    {:ok, %{client: client, first_message: first_message}}
+    {:ok, %{client: client, first_message: first_message, status: :open}}
   end
 
   @impl true
-  def disconnect(%{client: client}) do
+  def disconnect(%{client: client, status: :open}) do
     :ok = WebSocket.close(client)
   end
 
   @impl true
-  def refute_receive(%{client: client, first_message: first_message}) do
-    case first_message || WebSocket.recv(client, timeout: 100) do
-      {:ping, _} -> client.refute_receive(client)
-      {:ok, packet} -> raise "Unexpectedly received: #{inspect(packet)}"
-      {:error, _} -> {:ok, %{client: client, first_message: nil}}
+  def status(state)
+  def status(%{status: :closed} = state), do: {:closed, state}
+
+  def status(%{client: client} = state) do
+    # No idea why, but the first ping is always successful, regardless of whether the
+    # connection is still alive. As a workaround, we simply invoke ping twice.
+    case WebSocket.ping(client) do
+      {:error, _} ->
+        {:closed, %{state | status: :closed}}
+
+      _app_data ->
+        # The second ping is successful too when done right after the first one.. O_o
+        :timer.sleep(100)
+
+        case WebSocket.ping(client) do
+          {:error, _} -> {:closed, %{state | status: :closed}}
+          _app_data -> {:open, state}
+        end
     end
   end
 
   @impl true
-  def read_event(%{client: client, first_message: first_message}, event_type) do
+  def refute_receive(%{client: client, first_message: first_message, status: :open} = state) do
+    case first_message || WebSocket.recv(client, timeout: 100) do
+      {:ping, _} -> client.refute_receive(client)
+      {:ok, packet} -> raise "Unexpectedly received: #{inspect(packet)}"
+      {:error, _} -> {:ok, %{state | first_message: nil}}
+    end
+  end
+
+  @impl true
+  def read_event(
+        %{client: client, first_message: first_message, status: :open} = state,
+        event_type
+      ) do
     {:text, data} = first_message || WebSocket.recv!(client)
 
     cloud_event =
@@ -215,12 +267,21 @@ defmodule WsClient do
         %{"cloudEventsVersion" => "0.1", "eventType" => ^event_type} = cloud_event -> cloud_event
       end
 
-    {cloud_event, %{client: client, first_message: nil}}
+    {cloud_event, %{state | first_message: nil}}
   end
 
   @impl true
-  def read_welcome_event(client), do: read_event(client, "rig.connection.create")
+  def read_welcome_event(state), do: read_event(state, "rig.connection.create")
 
   @impl true
-  def read_subscriptions_set_event(client), do: read_event(client, "rig.subscriptions_set")
+  def read_subscriptions_set_event(state), do: read_event(state, "rig.subscriptions_set")
+
+  @impl true
+  def try_connect_then_disconnect(params \\ []) do
+    {:ok, client} = connect(params)
+    disconnect(client)
+    {:ok, client}
+  rescue
+    err in [TestClient.ConnectionError] -> {:error, err}
+  end
 end